requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.9
django-allauth==0.60.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
environs==10.0.0
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
marshmallow==3.20.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
pillow-heif==0.14.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wagtail==5.2.2
webencodings==0.5.1
Willow==1.6.3
